tests/wpt/web-platform-tests/permissions/test-periodic-background-sync-permission.html

<!doctype html>
<meta charset=utf-8>
<title>Test Periodic Background Sync Permission.</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>

<script>
  promise_test(async test => {
    const result = await navigator.permissions.query({name:'periodic-background-sync'});
    assert_true(result instanceof PermissionStatus);
  });
</script>
